ADP improvements #1

+++ com-mmxlabs-models-lng-transformer/com.mmxlabs.models.lng.transformer.ui/src/com/mmxlabs/models/lng/transformer/optimiser/lightweightscheduler/metaheuristic/SimpleLightweightSequenceOptimiser.java
@@ -43,7 +43,7 @@ public class SimpleLightweightSequenceOptimiser implements ILightWeightSequenceO
 	     List<Set<Integer>> cargoVesselRestrictions = lightWeightOptimisationData.getCargoVesselRestrictions();
 	     int[][][] cargoToCargoMinTravelTimes = lightWeightOptimisationData.getCargoToCargoMinTravelTimes();
 	     int[][] cargoMinTravelTimes = lightWeightOptimisationData.getCargoMinTravelTimes();
-	     double[] volumes = lightWeightOptimisationData.getCargoesVolumes();
+	     long[] volumes = lightWeightOptimisationData.getCargoesVolumes();
 	     LightWeightCargoDetails[] cargoDetails = lightWeightOptimisationData.getCargoDetails();
